Refactoring - changed name of DiffWindow into TextViewWindow
