+++ Application.h
@@ -10,7 +10,7 @@
 
 #include "AppMenu.h"
 #include "AppScreen.h"
-#include "DiffWindow.h"
+#include "TextViewWindow.h"
 #include "Command.h"
 
 class Application
@@ -32,8 +32,8 @@ private:
 
   bool m_bExitRequested;
   AppScreen* m_pScreen;
-  DiffWindow* m_pLeftWin;
-  DiffWindow* m_pRightWin;
+  TextViewWindow* m_pLeftWin;
+  TextViewWindow* m_pRightWin;
 
   AppMenu* m_pMenu;
   Command* m_pCmdOpenLeftFile;
